Move Factorial calculations to HomeLayout(Fragment)
