+++ app/src/main/java/com/solikil/kotlin/et2amel/app/mainScreen/features/homeFeature/view/HomeLayout.kt
@@ -5,9 +5,18 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.solikil.kotlin.et2amel.app.FactorialCalculator
+import com.solikil.kotlin.et2amel.app.NotificationUtil
 import com.solikil.kotlin.et2amel.app.R
+import kotlinx.android.synthetic.main.fragment_home.view.*
 
 class HomeLayout : Fragment() {
+
+    private val notificationUtil: NotificationUtil by lazy {
+        NotificationUtil(
+            this.requireActivity()
+        )
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -16,6 +25,21 @@ class HomeLayout : Fragment() {
     ): View? {
         val root = inflater.inflate(R.layout.fragment_home, container, false)
         // val textView: TextView = root.findViewById(text_home)
+        root.button_compute.setOnClickListener {
+            val input = root.edit_text_factorial.text.toString().toInt()
+            val result = FactorialCalculator.computeFactorial(
+                input
+            ).toString()
+
+            root.text_result.text = result
+            root.text_result.visibility = View.VISIBLE
+
+            notificationUtil.showNotification(
+                context = this.requireActivity(),
+                title = getString(R.string.notification_title),
+                message = result
+            )
+        }
 
         return root
     }
