+++ app/src/main/java/com/solikil/kotlin/et2amel/app/mainScreen/view/MainActivity.kt
@@ -4,19 +4,12 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.AppBarConfiguration
+import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.solikil.kotlin.et2amel.app.NotificationUtil
 import com.solikil.kotlin.et2amel.app.R
-import androidx.navigation.ui.setupActionBarWithNavController
 
 class MainActivity : AppCompatActivity() {
-
-    private val notificationUtil: NotificationUtil by lazy {
-        NotificationUtil(
-            this
-        )
-    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -28,25 +21,10 @@ class MainActivity : AppCompatActivity() {
             .findFragmentById(R.id.nav_host_fragment) as NavHostFragment
         val navController = navHostFragment.navController
 
-        val appBarConfiguration = AppBarConfiguration(setOf(
-            R.id.navigation_home, R.id.navigation_timer))
+        val appBarConfiguration =
+            AppBarConfiguration(setOf(R.id.navigation_home, R.id.navigation_timer))
         setupActionBarWithNavController(navController, appBarConfiguration)
+
         navView.setupWithNavController(navController)
-
-      /*  button_compute.setOnClickListener {
-            val input = edit_text_factorial.text.toString().toInt()
-            val result = FactorialCalculator.computeFactorial(
-                input
-            ).toString()
-
-            text_result.text = result
-            text_result.visibility = View.VISIBLE
-
-            notificationUtil.showNotification(
-                context = this,
-                title = getString(R.string.notification_title),
-                message = result
-            )
-        } */
     }
 }
